+++ src/main/java/de/theidler/create_mobile_packages/items/portable_stock_ticker/OpenPortableStockTicker.java
@@ -7,6 +7,9 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.SimpleMenuProvider;
+import net.minecraft.world.item.ItemStack;
+
+import static de.theidler.create_mobile_packages.items.portable_stock_ticker.LogisticallyLinkedItem.isTuned;
 
 public class OpenPortableStockTicker implements ServerboundPacketPayload {
     public static final OpenPortableStockTicker INSTANCE = new OpenPortableStockTicker();
@@ -17,6 +20,15 @@ public class OpenPortableStockTicker implements ServerboundPacketPayload {
 
     @Override
     public void handle(ServerPlayer player) {
+
+        ItemStack stack = PortableStockTicker.find(player.getInventory());
+        if (stack == null || !(stack.getItem() instanceof PortableStockTicker)) return;
+
+        if (!isTuned(stack)) {
+            player.displayClientMessage(Component.translatable("item.create_mobile_packages.portable_stock_ticker.not_linked"), true);
+            return;
+        }
+
         player.openMenu(new SimpleMenuProvider(
                 (id, inv, ply) -> new PortableStockTickerMenu(id, inv),
                 Component.translatable("item.create_mobile_packages.portable_stock_ticker")